inventary-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
grpcio
grpcio-tools
protobuf

